Trying to learn nextjs. Lol. Tough stuff! Testing a few changes at a time.
Trying to setup a navbar. While learning next & git branches.

This branch is my first go at trying to figure out the structure of a next app via trying to take it apart.

I plan to use this branch and it's commit history to track development in my learning how to manipulate app.

Once I figure out that part then I may start using a new branch to  build a specific part. Mostly I want another branch so that I can deepen understanding of merges and...
I'll probably try to create a merge conflict so I can learn that part.

Right here even, is a bit of a test; I'm testing to see how I want to use this type of notes.

May use this as a place to store the questions that I have to leave unanswered until next touch?

For example, looks like any file that I "save" is changing the single quotes to double quotes, is that convention on all nextjs projects?

I'm not sure how the notes are gonna look, and I have to leave that until next touch because I want to simulate real world scenarios.

Could this be a journal of sorts? It might even be designed for such, however I don't have any experience with the layout yet.